Refactor : club->favorite으로 getRecruitClubs 로직 이동
